+++ slack-plugin/src/main/kotlin/slack/gradle/SlackBasePlugin.kt
@@ -32,6 +32,7 @@ import org.gradle.kotlin.dsl.configure
 import org.gradle.kotlin.dsl.retry
 import org.gradle.kotlin.dsl.withType
 import org.jetbrains.kotlin.gradle.plugin.KotlinBasePlugin
+import slack.gradle.tasks.CoreBootstrapTask
 import slack.gradle.util.synchronousEnvProperty
 import slack.stats.ModuleStatsTasks
 
@@ -50,6 +51,7 @@ internal class SlackBasePlugin : Plugin<Project> {
         target.getVersionsCatalogOrNull() ?: error("SGP requires use of version catalogs!")
       StandardProjectConfigurations(slackProperties, versionCatalog).applyTo(target)
       target.configureTests(slackProperties)
+      CoreBootstrapTask.configureSubprojectBootstrapTasks(target)
 
       // Configure Gradle's test-retry plugin for insights on build scans on CI only
       // Thinking here is that we don't want them to retry when iterating since failure