+++ slack-plugin/src/main/kotlin/slack/gradle/tasks/BootstrapTask.kt
@@ -48,9 +48,11 @@ import org.gradle.jvm.toolchain.JavaLanguageVersion
 import org.gradle.jvm.toolchain.JavaLauncher
 import org.gradle.jvm.toolchain.JavaToolchainService
 import org.gradle.kotlin.dsl.listProperty
+import org.gradle.kotlin.dsl.named
 import org.gradle.kotlin.dsl.property
 import org.gradle.kotlin.dsl.register
 import org.gradle.kotlin.dsl.support.serviceOf
+import org.gradle.language.base.plugins.LifecycleBasePlugin
 import oshi.SystemInfo
 import slack.cli.AppleSiliconCompat
 import slack.cli.AppleSiliconCompat.isMacOS
@@ -308,20 +310,43 @@ constructor(objects: ObjectFactory, providers: ProviderFactory) : DefaultTask() 
   }
 
   public companion object {
+    private const val NAME: String = "bootstrap"
+
     /**
      * Current bootstrap version. Other bootstrap tasks can check against this as a minimum version.
      */
     public const val VERSION: Int = 2
 
+    internal fun isBootstrapEnabled(project: Project): Boolean {
+      return project.gradle.startParameter.taskNames.any { it == NAME }
+    }
+
+    internal fun configureSubprojectBootstrapTasks(project: Project) {
+      if (!isBootstrapEnabled(project)) return
+      val rootTask = project.rootProject.tasks.named<CoreBootstrapTask>(NAME)
+      // Clever trick to make this finalized by all bootstrap tasks and all other tasks depend on
+      // this, so bootstrap always runs first.
+      project.tasks.configureEach {
+        val task = this
+        if (name == NAME) return@configureEach
+        if (name == LifecycleBasePlugin.CLEAN_TASK_NAME) return@configureEach
+        if (this is BootstrapTask) {
+          rootTask.configure { finalizedBy(task) }
+        } else {
+          dependsOn(rootTask)
+        }
+      }
+    }
+
     public fun register(project: Project): TaskProvider<CoreBootstrapTask> {
       check(project.isRootProject) { "Bootstrap can only be applied to the root project" }
       val bootstrap =
-        project.tasks.register<CoreBootstrapTask>("bootstrap") {
+        project.tasks.register<CoreBootstrapTask>(NAME) {
           val jdkVersion = project.jdkVersion()
           val service = project.serviceOf<JavaToolchainService>()
           val defaultLauncher =
             service.launcherFor { languageVersion.set(JavaLanguageVersion.of(jdkVersion)) }
-          @Suppress("LeakingThis") launcher.convention(defaultLauncher)
+          this.launcher.convention(defaultLauncher)
           this.jdkVersion.set(jdkVersion)
 
           val cacheDirProvider = project.layout.projectDirectory.dir(".cache")
@@ -360,23 +385,6 @@ constructor(objects: ObjectFactory, providers: ProviderFactory) : DefaultTask() 
               .orElse(APPEND)
           )
         }
-
-      // Clever trick to make this finalized by all bootstrap tasks and all other tasks depend on
-      // this,
-      // so this always runs first.
-      // Safe to use allprojects because we only run this conditionally
-      project.allprojects {
-        tasks.configureEach {
-          val task = this
-          if (name == "bootstrap") return@configureEach
-          if (name == "clean") return@configureEach
-          if (this is BootstrapTask) {
-            bootstrap.configure { finalizedBy(task) }
-          } else {
-            dependsOn(bootstrap)
-          }
-        }
-      }
 
       return bootstrap
     }